+++ lib/events.php
@@ -1,29 +1,45 @@
 <?php
 
+/**
+ * This events class will be used in a future
+ * version of Kirby to provide event hooks in all kinds of places
+ * 
+ * @package Kirby CMS
+ */
 class KirbyEvents {
 
-  static $events = array();
+  // array with all collected events
+  static protected $events = array();
 
+  /**
+   * Registers a new event. 
+   * 
+   * @param string $events The name of the event
+   * @param func $callback The callback function
+   */
   static public function on($event, $callback) {
-    if(!isset(static::$events[$event])) static::$events[$event] = array();
+    if(!isset(self::$events[$event])) self::$events[$event] = array();
     if(is_array($callback)) {
       // attach all passed events at once
-      static::$events[$event] = array_merge(static::$events[$event], $callback);
+      self::$events[$event] = array_merge(self::$events[$event], $callback);
     } else {
       // attach a single new event
-      static::$events[$event][] = $callback;
+      self::$events[$event][] = $callback;
     }
   }
 
-  public function trigger($event) {
-    if(!empty(static::$events[$event])) {
-      foreach(static::$events[$event] as $callback) {
-        if(is_string($callback) && method_exists($this, $callback)) {
-          return $this->$callback();
-        } else if(is_callable($callback)) {
-          return $callback($this); 
-        }
-      } 
+  /**
+   * Triggers all available events for a given key
+   * 
+   * @param string $event The name of the event, that should be triggered
+   * @param array $arguments An optional array of arguments, which should be passed to the event
+   */
+  public function trigger($event, $arguments = array()) {
+    if(empty(self::$events[$event])) return false;
+    
+    foreach(self::$events[$event] as $callback) {
+      if(!is_callable($callback)) continue;
+      call_user_func_array($callback, $arguments);
     }
   }